Update comment and change some doubles to longs

+++ polaris-service/src/main/java/org/apache/polaris/service/ratelimiter/DefaultRateLimiterFactory.java
@@ -30,10 +30,10 @@ import java.util.concurrent.Future;
 @JsonTypeName("default")
 public class DefaultRateLimiterFactory implements RateLimiterFactory {
   @JsonProperty("requestsPerSecond")
-  private double requestsPerSecond;
+  private long requestsPerSecond;
 
   @JsonProperty("windowSeconds")
-  private double windowSeconds;
+  private long windowSeconds;
 
   @Override
   public Future<RateLimiter> createRateLimiter(String key) {
